Add mocked aio_s3fs fixture, skips if s3fs is not installed
