+++ pytest_aiomoto/plugin.py
@@ -4,6 +4,7 @@ pytest_plugins = [
     "pytest_aiomoto.aiomoto_fixtures",
     "pytest_aiomoto.aiomoto_lambda",
     "pytest_aiomoto.aiomoto_s3",
+    "pytest_aiomoto.aiomoto_s3fs",
 ]
